+++ src/input_context/input_condition/block_by.rs
@@ -39,7 +39,7 @@ impl<A: InputAction> InputCondition for BlockBy<A> {
         _value: ActionValue,
     ) -> ActionState {
         if let Some(action) = action_map.get(&TypeId::of::<A>()) {
-            if action.state() == ActionState::Fired {
+            if action.state == ActionState::Fired {
                 return ActionState::None;
             }
         } else {